Added Remember-Me Login, Long press to add tree pin
Saved tokens to phone so that you will be automatically signed in when
opening the app.
User is able to see if they are admin.
Added the ability to choose either tree or bird on the menu bar.
A long press on the map will display a dialog box which asks the user if
they want to add a tree.
A tree pin is added to map after selecting yes on dialog box